Send the result of summarization to the users
