+++ apps/agents/delivery/package.json
@@ -4,8 +4,8 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "dev": "dotenv -e ../../../packages/env/.env.agents.delivery -- bun run --hot src/index.ts",
-    "build": "bun build src/index.ts --outfile=./dist/server.js --target=bun"
+    "dev": "dotenv -e ../../../packages/env/.env.agents.delivery -- bun run --hot src/_main.ts",
+    "build": "bun build src/_main.ts --outfile=./dist/server.js --target=bun"
   },
   "dependencies": {
     "@workspace/agent-utils": "workspace:*",
@@ -17,6 +17,7 @@
   },
   "devDependencies": {
     "@types/bun": "latest",
+    "@workspace/typescript-config": "workspace:*",
     "dotenv-cli": "catalog:"
   }
 }
