+++ apps/agents/delivery/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "@workspace/typescript-config/node16.json",
+  "include": ["src"],
+  "exclude": ["node_modules"]
+}
